move helpers into CI and share them with non-CI
